penambahan otorisasi step dan ingredient

+++ app/Http/Controllers/IngredientController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Models\Ingredient;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Validator;
 
 class IngredientController extends Controller
@@ -13,9 +15,11 @@ class IngredientController extends Controller
      * Menyimpan data ingredient
      */
     public function create(Request $request){
+        if (! Gate::allows('create-ingredient', Session::get('recipe_id_r'))) {
+            abort(403);
+        }
         $validator = Validator::make($request->all(), [
             'value' => 'required',
-            'recipe_id' => 'required',
         ]);
         
         if ($validator->fails()) {
@@ -23,7 +27,7 @@ class IngredientController extends Controller
         }
         $ingredient = new Ingredient;
         $ingredient->value = $request->value;
-        $ingredient->recipe_id = $request->recipe_id;
+        $ingredient->recipe_id = Session::get('recipe_id_r');
         $ingredient->save();
         return back();
     }
@@ -41,8 +45,11 @@ class IngredientController extends Controller
         }
     
         $ingredient = Ingredient::find($id);
+        
+        if (! Gate::allows('admin-ingredient', $ingredient)) {
+            abort(403);
+        }
         $ingredient->value = $request->value;
-        $ingredient->recipe_id = $request->recipe_id;
         $ingredient->save();
         return back();
     }
@@ -50,6 +57,9 @@ class IngredientController extends Controller
      * Meghapus data ingredient
      */
     public function delete(Ingredient $ingredient){
+        if (! Gate::allows('admin-ingredient', $ingredient)) {
+            abort(403);
+        }
         $ingredient->delete();
         return back();
     }
